requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.1
requests==2.31.0
celery==5.3.4
redis==5.0.1
